Get started rewriting the primary Makefile
* Use <file>.<proc>.c for all files
* Assume .arm if .<proc> is left off
* Enable 'make <file>.<proc>.out' to define build target

Working toward supporting local Makefiles that define targets and use 'make -C /path/to/this/Makefile' to perform compilation simply.
