+++ .c9/runners/c.run
@@ -1,20 +1,13 @@
 // Create a custom Cloud9 runner - similar to the Sublime build system
-// For more information see http://docs.c9.io:8080/#!/api/run-method-run
+// For more information see https://docs.c9.io/docs/custom-runners
 {
     "script": [
-        "base=$file_base_name",
-// Find which model (Black, AI, Pocket, etc.) is running
-        "export MODEL=$(awk '{print $NF}' /proc/device-tree/model)",
-        "export TARGET=${base%%.*}",    
-// The .c file should have PRUN #defined.  It's read here
-        "export   PRUN=`awk '$1 ~ /^#define/ && $2 ~ /PRUN/ {print $3}' $file_name`",    
-        // "PRUN=${PRUN#pru*}",            // Remove'pru' from the name
-        // "echo base=$base, TARGET=$TARGET",
-        // "echo prun=$PRUN",
-        // "echo MODEL=$MODEL", 
-        "cd $file_path",
-        "make $args"
-        ],
+        "export BASE=${file_base_name%%.*}",    // Extract everything before the first '.'
+        "export PROC=$(echo $file_base_name | sed 's/$BASE.//')",
+        "export MODEL=$(cat /proc/device-tree/model | sed 's/ /_/g' | tr -d '\000')",
+            // Use device-tree model for board name with spaces turned to underscores and clean termination
+        "make -C /var/lib/cloud9/Makefile $args"
+    ],
 
 //    "info" : "Started $project_path: $file_name\n\nFile path: $file_path\nFile: $file\nArguments: $args\nFile name: $file_name\nFile extension: $file_extension\nFile base name: $file_base_name\nPackages: $packages\nProject path: $project_path\nProject name: $project_name\nHostname: $hostname\nHostname path: $hostname_path\nURL: $url\nPort: $port\nIP: $ip\nCommand: $command\nPython: $python\nPython path: $python_path",
     "env" : {},
