+++ BeagleBone/AI/pru/init_pins_empty.h
@@ -1,5 +1,4 @@
-// Defins an empty table to write_init_pins wont' complain
-
+// Define an empty table to write_init_pins won't complain
 #pragma DATA_SECTION(init_pins, ".init_pins")
 #pragma RETAIN(init_pins)
 const char init_pins[] =
